Unit Tests, and bitrate control, TravisCI
A lot went into this one, most unit testing is done outside of the middleware and a few integration tests. Bitrate control was integrated, a config manager class is now used to simplify config variables. Each path is split into modules for simplicity and readability. TravisCI is being actively implemented.